Onix/Customers/Retail/2025/business-analysis-platform/requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
plotly==5.18.0
pyyaml==6.0.1
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.2
scipy==1.11.3
statsmodels==0.14.0

# API dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-watch==4.2.0
httpx==0.25.2

